requirements.txt
streamlit
torch
timm
scikit-learn
joblib
Pillow
torchvision